+++ endpoints.go
@@ -7,29 +7,29 @@ type agentEndpoints struct {
 	me  string
 }
 type articleEndpoints struct {
-	delete func(int) string
-	get    func(int) string
+	delete func(int64) string
+	get    func(int64) string
 }
 
 type categoryEndpoints struct {
-	folders func(int) string
+	folders func(int64) string
 }
 
 type companyEndpoints struct {
 	all    string
 	create string
-	update func(int) string
+	update func(int64) string
 }
 
 type contactEndpoints struct {
 	all    string
 	create string
-	update func(int) string
+	update func(int64) string
 	search func(string) string
 }
 
 type folderEndpoints struct {
-	articles func(int) string
+	articles func(int64) string
 }
 
 type groupEndpoints struct {
@@ -38,7 +38,7 @@ type groupEndpoints struct {
 
 type slaPolicyEndpoints struct {
 	all    string
-	update func(int) string
+	update func(int64) string
 }
 
 type solutionEndpoints struct {
@@ -51,10 +51,10 @@ type solutionEndpoints struct {
 type ticketEndpoints struct {
 	all             string
 	create          string
-	view            func(int) string
+	view            func(int64) string
 	search          func(string) string
-	reply           func(int) string
-	conversations   func(int) string
+	reply           func(int64) string
+	conversations   func(int64) string
 	updatedSinceAll func(string) string
 }
 
@@ -74,12 +74,12 @@ var endpoints = struct {
 	companies: companyEndpoints{
 		all:    "/api/v2/companies",
 		create: "/api/v2/companies",
-		update: func(id int) string { return fmt.Sprintf("/api/v2/companies/%d", id) },
+		update: func(id int64) string { return fmt.Sprintf("/api/v2/companies/%d", id) },
 	},
 	contacts: contactEndpoints{
 		all:    "/api/v2/contacts",
 		create: "/api/v2/contacts",
-		update: func(id int) string { return fmt.Sprintf("/api/v2/contacts/%d", id) },
+		update: func(id int64) string { return fmt.Sprintf("/api/v2/contacts/%d", id) },
 		search: func(query string) string { return fmt.Sprintf("/api/v2/search/contacts?%s", query) },
 	},
 	groups: groupEndpoints{
@@ -87,28 +87,28 @@ var endpoints = struct {
 	},
 	slaPolicies: slaPolicyEndpoints{
 		all:    "/api/v2/sla_policies",
-		update: func(id int) string { return fmt.Sprintf("/api/v2/sla_policies/%d", id) },
+		update: func(id int64) string { return fmt.Sprintf("/api/v2/sla_policies/%d", id) },
 	},
 	solutions: solutionEndpoints{
 		categories: "/api/v2/solutions/categories",
 		category: categoryEndpoints{
-			folders: func(id int) string { return fmt.Sprintf("/api/v2/solutions/categories/%d/folders", id) },
+			folders: func(id int64) string { return fmt.Sprintf("/api/v2/solutions/categories/%d/folders", id) },
 		},
 		folder: folderEndpoints{
-			articles: func(id int) string { return fmt.Sprintf("/api/v2/solutions/folders/%d/articles", id) },
+			articles: func(id int64) string { return fmt.Sprintf("/api/v2/solutions/folders/%d/articles", id) },
 		},
 		articles: articleEndpoints{
-			delete: func(id int) string { return fmt.Sprintf("/api/v2/solutions/articles/%d", id) },
-			get:    func(id int) string { return fmt.Sprintf("/api/v2/solutions/articles/%d", id) }, // Not currently in use
+			delete: func(id int64) string { return fmt.Sprintf("/api/v2/solutions/articles/%d", id) },
+			get:    func(id int64) string { return fmt.Sprintf("/api/v2/solutions/articles/%d", id) }, // Not currently in use
 		},
 	},
 	tickets: ticketEndpoints{
 		all:           "/api/v2/tickets",
 		create:        "/api/v2/tickets",
-		view:          func(id int) string { return fmt.Sprintf("/api/v2/tickets/%d", id) },
+		view:          func(id int64) string { return fmt.Sprintf("/api/v2/tickets/%d", id) },
 		search:        func(query string) string { return fmt.Sprintf("/api/v2/search/tickets?%s", query) },
-		reply:         func(id int) string { return fmt.Sprintf("/api/v2/tickets/%d/reply", id) },
-		conversations: func(id int) string { return fmt.Sprintf("/api/v2/tickets/%d/conversations", id) },
+		reply:         func(id int64) string { return fmt.Sprintf("/api/v2/tickets/%d/reply", id) },
+		conversations: func(id int64) string { return fmt.Sprintf("/api/v2/tickets/%d/conversations", id) },
 		updatedSinceAll: func(timeString string) string {
 			return fmt.Sprintf("/api/v2/tickets?updated_since=%s", timeString)
 		},
